chore: move colors to shared
