--- src/themes/colors/index.ts
@@ -1,1 +1,0 @@
-export { colors } from './colors';
